requirements.txt
appnope==0.1.4
asttokens==2.4.1
attrs==23.2.0
certifi==2024.6.2
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
comm==0.2.2
contourpy==1.2.1
cycler==0.12.1
debugpy==1.8.1
decorator==5.1.1
executing==2.0.1
fiona==1.9.6
fonttools==4.51.0
geopandas==0.14.4
ipykernel==6.29.4
ipython==8.24.0
jedi==0.19.1
joblib==1.4.2
jupyter_client==8.6.1
jupyter_core==5.7.2
kiwisolver==1.4.5
llvmlite==0.42.0
matplotlib==3.8.4
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
numba==0.59.1
numpy==1.26.4
packaging==24.0
pandas==2.2.2
parso==0.8.4
patsy==0.5.6
pexpect==4.9.0
pillow==10.3.0
platformdirs==4.2.1
plotly==5.22.0
plotly-express==0.4.1
prompt-toolkit==3.0.43
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.18.0
pynndescent==0.5.12
pyparsing==3.1.2
pyproj==3.6.1
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
pyzmq==26.0.3
scikit-learn==1.4.2
scipy==1.13.0
seaborn==0.13.2
shapely==2.0.4
six==1.16.0
stack-data==0.6.3
statsmodels==0.14.2
tenacity==8.3.0
threadpoolctl==3.5.0
tornado==6.4
tqdm==4.66.4
traitlets==5.14.3
typing_extensions==4.11.0
tzdata==2024.1
umap-learn==0.5.6
wcwidth==0.2.13
xgboost==2.0.3
